# openapi.yaml
# ===========================================================================
# 纯白图识别 API — OpenAPI 3.0 Schema
# ===========================================================================
# 用途：直接导入 Coze 插件 / Dify 自定义工具
# 使用方法：
#   1. 将 servers.url 替换为你的公域地址（如 ngrok 链接）
#   2. 在 Coze：插件 → 创建插件 → 导入 OpenAPI Schema
#   3. 在 Dify：工具 → 自定义 → 创建 → 导入 OpenAPI Schema
# ===========================================================================

openapi: 3.0.0
info:
  title: 纯白图识别 API
  description: |
    基于像素分析的纯白图 / 空图识别服务。

    **不依赖任何 AI 模型**，纯 NumPy + Pillow 像素分析，毫秒级响应，
    可部署到任意云服务器 / 容器，适合批量图片质检、电商主图巡检等场景。

    判定结果：
    - `is_white_image = true` → 纯白图 / 纯色空图（应标记或退回）
    - `is_white_image = false` → 正常图（含商品主体等真实内容）

    三种输入方式：文件上传、URL、Base64。
  version: 1.0.0

servers:
  - url: https://your-ngrok-url.ngrok-free.dev
    description: 替换为你的公域服务地址（ngrok / 云服务器域名）

paths:
  /api/v1/detect:
    post:
      summary: 纯白图识别（文件上传）
      description: 上传图片文件，识别是否为纯白图 / 空图
      operationId: detectByFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 图片文件（png/jpg/webp，最大 20MB）
                white_threshold:
                  type: number
                  minimum: 0.5
                  maximum: 1.0
                  default: 0.99
                  description: 近白像素占比阈值，≥ 此值判为纯白图
                max_color_variety:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 3
                  description: 颜色种类上限，≤ 此值且整体浅色判为纯色空图
      responses:
        "200":
          description: 识别成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectResponse"

  /api/v1/detect/url:
    post:
      summary: 纯白图识别（URL 输入）
      description: 通过图片 URL 进行纯白图识别
      operationId: detectByUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_url]
              properties:
                image_url:
                  type: string
                  format: uri
                  description: 图片 URL 地址
                white_threshold:
                  type: number
                  minimum: 0.5
                  maximum: 1.0
                  default: 0.99
                max_color_variety:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 3
      responses:
        "200":
          description: 识别成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectResponse"

  /api/v1/detect/base64:
    post:
      summary: 纯白图识别（Base64 输入）
      description: 通过 Base64 编码的图片进行纯白图识别
      operationId: detectByBase64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_base64]
              properties:
                image_base64:
                  type: string
                  description: Base64 编码的图片数据（支持 data:image/xxx;base64, 前缀）
                white_threshold:
                  type: number
                  minimum: 0.5
                  maximum: 1.0
                  default: 0.99
                max_color_variety:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 3
      responses:
        "200":
          description: 识别成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectResponse"

  /api/v1/health:
    get:
      summary: 健康检查
      description: 检查服务状态
      operationId: healthCheck
      responses:
        "200":
          description: 服务状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  service:
                    type: string
                  version:
                    type: string
                  engine:
                    type: string
                  model_required:
                    type: boolean

components:
  schemas:
    DetectResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否识别成功
        is_white_image:
          type: boolean
          description: "true=纯白图/纯色空图（应标记），false=正常图"
        category:
          type: string
          description: 分类（纯白图 / 纯色图 / 正常图）
        white_ratio:
          type: number
          description: 近白像素占比（0~1）
        color_variety:
          type: integer
          description: 量化后颜色种类数
        dominant_color:
          type: array
          items:
            type: integer
          description: 主色 RGB
        image_size:
          type: array
          items:
            type: integer
          description: 图片尺寸 [宽, 高]
        format:
          type: string
          description: 图片格式（JPEG/PNG/...）
        confidence:
          type: number
          description: 判定置信度（0~1）
        reason:
          type: string
          description: 判定原因说明
        processing_time:
          type: number
          description: 处理耗时（秒）
        error:
          type: string
          description: 失败时的错误码